#102 #99 Ignoring and clearing out old temporary LESS files
These files are used to implement the LESS globals system and should normally be deleted as soon as they are used. However, some situations lead to them sticking around. This change ensures that they are not subsequently processed as files to be included and that we attempt to clear them out before processing all the LESS includes. This is achieved by using a obscure naming pattern for the temporary files.